backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "PyJWT>=2.9",
    "httpx>=0.27",
    "pytest>=8.0",
]
